Bug Fixed for DAO clasess

+++ MyProject/src/main/java/com/nik/college/dao/CourseDao.java
@@ -4,11 +4,11 @@ import com.nik.college.domain.Course;
 
 public interface CourseDao {
 
-	public Course findCourseByCourseCode(long id);
+	public Course findCourseByCourseCode(int id);
 	public Course findCourseByName(String name);
 	public void insertCourse(Course course);
 	public void deleteCourse(Course course);
-	public int updateCourse(long id, Course course);
+	public int updateCourse(int id, Course course);
 	public int getCourseCount();
 	
 }
