+++ MyProject/src/main/java/com/nik/college/service/CourseDaoService.java
@@ -3,11 +3,11 @@ package com.nik.college.service;
 import com.nik.college.domain.Course;
 
 public interface CourseDaoService {
-	public Course findCourseByCourseCode(long id);
+	public Course findCourseByCourseCode(int id);
 	public Course findCourseByName(String name);
 	public void insertCourse(Course course);
 	public void deleteCourse(Course course);
-	public int updateCourse(long id, Course course);
+	public int updateCourse(int id, Course course);
 	public int getCourseCount();
 
 }
